pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "signal_backend"
version = "0.1.0"
description = "Dataset ingestion, validation, split, and training utilities for Signal backend."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "joblib>=1.5.0",
    "pandas>=2.2.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.4.0",
    "torch>=2.2.0",
    "transformers>=4.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]